src/ml-training/requirements.txt
# ml-training/requirements.txt
celery==5.3.1
flower==2.0.1
redis==4.6.0
xgboost==1.7.6
scikit-learn==1.3.0
imbalanced-learn==0.11.0
pandas==2.0.3
numpy==1.24.3
psycopg2-binary==2.9.7
python-dotenv==1.0.0
loguru==0.7.0
joblib==1.3.2